requirements.txt
azure-functions
playwright
azure-storage-blob
